Restructuring for easier understanding
