+++ src/Proto.Cluster/IClusterProvider.cs
